[SPARK-50579][SQL] Fix `truncatedString`
### What changes were proposed in this pull request?
In the PR, I propose to respect the `maxFields` argument of `SparkStringUtils.truncatedString` in all cases independently from the size of the another argument `seq`. If size of `seq` is grater or equal to `maxFields`, output exact `maxFields` elements of `seq`.

Also, if the number of printed elements is zero, don't print the redundant comma `,` like `[, ... 100 more fields]`, and don't overflow when `maxFields` is `Int.MinValue`.

### Why are the changes needed?
To make output consistent for the same `maxFields`. For example, if `maxFields` is 2:
```scala
truncatedString(Seq(1, 2), "[", ", ", "]", maxFields = 2) -> "[1, 2]"
```
but for more elements in the input sequence, it prints only one element:
```scala
truncatedString(Seq(1, 2, 3), "[", ", ", "]", maxFields = 2) -> "[1, ... 2 more fields]"
```
though the expected output is **"[1, 2, ... 1 more fields]"**.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
By existing tests from `UtilSuite` that are moved to `StringUtilsSuite` and additional checks:
```
$ build/sbt "test:testOnly *StringUtilsSuite"
```

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #49187 from MaxGekk/fix-truncatedString.

Authored-by: Max Gekk <max.gekk@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>