[SPARK-50597][SQL] Refactor batch construction in Optimizer.scala and SparkOptimizer.scala
### What changes were proposed in this pull request?

See description. Previously, it was a pain to reorder batches and guard behavior of certain batches / sequences of batches by a flag. This was primarily due to ample usage of `::`, `:::`, and `:+` to juggle rules and batches around which imposed syntactic limitations.

After this change, we keep a single sequence `allBatches`, that can contain either `Batch` or `Seq[Batch]` elements to allow further groupings, which is later flattened into a single `Seq[Batch]`.

We avoid any usage of `::`, `:::`, and `:+`.

To add/replace a flag-guarded batch of sequence of batches, write a function that returns either `Batch` of `Seq[Batch]` with desired behavior, and add/replace in the relevant place in the `allBatches` list.

### Why are the changes needed?

This simplifies further restructuring and reordering of batches.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

No tests.

### Was this patch authored or co-authored using generative AI tooling?

Closes #49208 from anton5798/batch-refactor.

Authored-by: Anton Lykov <anton.lykov@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>