[SPARK-50472][SQL] Introduce initial implementation of the single-pass Analyzer
### What changes were proposed in this pull request?

This is the initial PR for the single-pass Analyzer project.

Coverage:
- Scans of csv/json/parquet/orc/hive tables and files
- `VALUES`
- `WHERE`
- `LIMIT`
- Trivial star expansion
- Literals
- Aliases
- Binary arithmetic
- Binary logic

Main components:

`HybridAnalyzer`:
The analysis is now routed between the fixed-point and single-pass by the `HybridAnalyzer`. There are 3 modes:
- Fixed-point. In that case `HybridAnalyzer` just calls `executeAndTrack`. This is the default one.
- Dual. This one is activated by the `spark.sql.analyzer.singlePassResolver.dualRunWithLegacy` flag and involves a filtering logic based on the `ResolverGuard`. First, the query is always analyzed by the fixed-point `Analyzer`, and then, if it's supported by the single-pass, it's analyzed by the `Resolver`.
- Single-pass. In that case `HybridAnalyzer` calls the single-pass `Resolver` without any filtering or routing. This is useful for debugging.

`ResolverGuard`:
This component allows `HybridAnalyzer` to figure out whether a SQL query/DataFrame program is supported by the single-pass `Resolver`. Here we do a simple pass over the operator/expression tree with a short-circuited boolean logic.

`Resolver`:
This is the main entry point for the single-pass analysis. All the operators are resolved here. If the new operator is introduced into the Catalyst, it should be handled in the `resolve(...)` pattern match. The operator is either resolved here, or the resolution is delegated to the extensions.

`ExpressionResolver`:
Similar to the `Resolver`, this is the main entry point for the expression analysis. The resolution either happens here, or is delegated to specific resolvers (e.g. `BinaryArithmeticResolver`, `TimeZoneAwareExpressionResolver`). The `Resolver` calls the `ExpressionResolver` to resolve expressions nested in operators.

`MetadataResolver`:
The `MetadataResolver` is used by the `Resolver` to look up relation metadata and store it in a map by `RelationId`. There are two modes:
- `AnalyzerBridgeState` from the fixed-point `Analyzer` is available. In that case we just reuse the existing relation metadata from the fixed-point run. This is done to avoid latency/consistency issues.
- Otherwise, we collect the metadata using `RelationResolution`.

`ResolverExtension`:
This is a trait that allows extending the `Resolver`. Its `resolveOperator` must deterministically match the unresolved plan node - a conflict between the configured extensions is disallowed.

`NameScope`:
This component controls the visibility of names (tables, columns, aliases), star expansion for tables and structs and operator output handling. A new `NameScope` is created when a new scope is introduced (`Project`, `Aggregate`). We add operator output to the `NameScope` to later look up unresolved attributes and aliases. The `NameScope`s are stacked together to further support correlated names.

`DataSourceResolver`/`DataSourceWithHiveResolver`:
Those extensions allow us to resolve relation nodes (e.g. `UnresolvedCatalogRelation`, `LogicalRelation`, `HiveTableRelation`) defined outside of `catalyst` package to avoid depending on those packages (that would require a cyclic import).

`ResolutionValidator`/`ExpressionResolutionValidator`:
These components work together on the resolved tree and assert invariants that should be true no matter which SQL query/DataFrame program was passed to the single-pass analyzer. For example, we validate that expressions used in project lists actually exist in the lower operator outputs (using `AttributeScopeStack`). Every `resolve*` method in `Resolver`/`ExpressionResolver` must have its `validate*` counterpart in the `ResolutionValidator`/`ExpressionResolutionValidator`.

`TracksResolvedNodes`:
This is another validation mechanism which asserts that the analysis is actually single-pass - we store a reference to any resolved node in a set, and prior to the resolution check that a reference to the passed node hasn't been previously analyzed.

Thanks mihailoale-db and mihailotim-db for working on that!

### Why are the changes needed?

We want to build a new single-pass Analyzer to replace the fixed-point one.

SPIP Jira: https://issues.apache.org/jira/browse/SPARK-49834
SPIP: https://docs.google.com/document/d/1dWxvrJV-0joGdLtWbvJ0uNyTocDMJ90rPRNWa4T56Og

### Does this PR introduce _any_ user-facing change?

No, the single-pass Analyzer is disabled by default (under the `spark.sql.analyzer.singlePassResolver.enabled` and `spark.sql.analyzer.singlePassResolver.dualRunWithLegacy` flags).

### How was this patch tested?

- New test suites
- Ran some suites with `spark.sql.analyzer.singlePassResolver.dualRunWithLegacy` enabled

### Was this patch authored or co-authored using generative AI tooling?

Copilot.

Closes #49029 from vladimirg-db/vladimirg-db/single-pass-analyzer/initial-state.

Authored-by: Vladimir Golubev <vladimir.golubev@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>