[SPARK-49632][SQL] Remove the ANSI config suggestion in CANNOT_PARSE_TIMESTAMP
### What changes were proposed in this pull request?
This PR changes the message returned on a failure of `ToTimestamp` family of expressions.

### Why are the changes needed?
CANNOT_PARSE_TIMESTAMP contains suggested fix for turning off ANSI mode. Now that in Spark 4.0.0 we have moved to ANSI mode on by default, we want to keep suggestions of this kind to the minimum.

### Does this PR introduce _any_ user-facing change?
Yes, error message is changing.

### How was this patch tested?
Existing tests span the error message change.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #49227 from mihailom-db/cannot_parse_timestamp.

Authored-by: Mihailo Milosevic <mihailo.milosevic@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>