[SPARK-50558][SQL] Add configurable logging limits for the number of elements in InSet and In
### What changes were proposed in this pull request?

Introduce a new SQL conf `OPTIMIZER_INSET_MEMBER_LOGGING_LIMIT` that allows to limit the number of elements logged for In and InSet in toString. The current implementation does log arbitrarily many elements and in the case of InSet even sorts them before, which can lead to a lot of unwanted log volumes and have significant performance impact.
If clients require all elements, they can set to conf to a very high value or preferably use the `sql` function.

### Why are the changes needed?

I have observed performance impact of the logging

### Does this PR introduce _any_ user-facing change?

Teh logging string changes. Given that there is also the `sql` function, clients should not depend on the output of the logging function.

### How was this patch tested?

New tests have been added

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #49177 from olaky/spark-50558-do-not-log-all-values-in-in-set-expressions.

Authored-by: Ole Sasse <ole.sasse@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>