[SPARK-50612][SQL] Normalize ordering in the project list of an inner Project
### What changes were proposed in this pull request?
Normalize ordering in the project list of an inner Project. For example, following two plans should be the same after normalization:
```
Project[col1]                                  Project[col1]
:- Project[col1, col2]                         :- Project[col2, col1]
```
This is because only the top most ordering matters for comparing plans and inner project list can be computed in a non-deterministic way (i.e. via `HashSet`/`HashMap`)

### Why are the changes needed?

For plan comparison in the single-pass resolver.

### Does this PR introduce _any_ user-facing change?

### How was this patch tested?
Added a new test suite.

### Was this patch authored or co-authored using generative AI tooling?

Closes #49230 from mihailotim-db/mihailotim-db/normalize_plan.

Authored-by: Mihailo Timotic <mihailo.timotic@databricks.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>