[SPARK-50573][SS] Adding State Schema ID to State Rows to schema evolution
### What changes were proposed in this pull request?

In order to support schema evolution without a full rewrite, we are prepending each state row with a state schema ID, which we will use in subsequent changes to look up the schema that a row was written with. This change adds state schema IDs to both key and value rows in order to enable schema evolution when Avro is enabled.
For NoPrefixKeys and Values, the state schema ID is prepended to the row.
For PrefixKeys and RangeKeys, the state schema ID is prepended to the remaining portion of the key (so the format is prefix key, schema ID, remaining key)
### Why are the changes needed?

These changes are needed in order to support schema evolution when Avro is enabled.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Added unit test in RocksDBSuite to test all 4 prefix behaviors.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #49171 from ericm-db/ssm-temp.

Authored-by: Eric Marnadi <eric.marnadi@databricks.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>