[SPARK-50681][PYTHON][CONNECT] Cache the parsed schema for MapInXXX and ApplyInXXX
### What changes were proposed in this pull request?
Cache the parsed schema for MapInXXX and ApplyInXXX

### Why are the changes needed?
The specified schema for MapInXXX and ApplyInXXX has been cached when the schema is `StructType`.

For a `str` schema, it is parsed in `UserDefinedFunction.returnType` but will not be cached.
In this case, we can move this parse ahead so it can be cached in MapInXXX and ApplyInXXX

### Does this PR introduce _any_ user-facing change?
no

### How was this patch tested?
added test

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #49305 from zhengruifeng/py_cache_ddl.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
Signed-off-by: Ruifeng Zheng <ruifengz@apache.org>