[SPARK-50682][SQL] Inner Alias should be canonicalized
### What changes were proposed in this pull request?

This PR fixes plan canonicalization to normalize the expr ID of inner Aliases to 0. We should not assume that all inner Aliases will be removed by the analyzer, as Alias with metadata won't be removed.

### Why are the changes needed?

To avoid cache misses that should not happen.

### Does this PR introduce _any_ user-facing change?

no, table cache is a perf feature.

### How was this patch tested?

a new test

### Was this patch authored or co-authored using generative AI tooling?

no

Closes #49309 from cloud-fan/alias.

Authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>