[SPARK-50614][SQL] Fix shredding debug message
### What changes were proposed in this pull request?

A debug message added in https://github.com/apache/spark/pull/49234 was missing a brace. As a result, the message was printing the non-pretty string representation of the struct, followed by the string `.prettyJson}`. This PR fixes it.

### Why are the changes needed?

Cleaner debug messages.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Manually verified in spark-shell that the message looks wrong without the change, and correct with the change.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #49345 from cashmand/fix_debug_msg.

Authored-by: cashmand <david.cashman@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>