[SPARK-50734][SQL] Add catalog API for creating and registering SQL UDFs
### What changes were proposed in this pull request?

This PR adds catalog APIs to support the creation and registration of SQL UDFs. It uses Hive Metastore to persist a SQL UDF by deserializing the function information into a FunctionResource and storing it in Hive (toCatalogFunction). During resolution, it retrieves the catalog function and deserializes it into a SQLFunction.

This PR only adds the catalog API, and a subsequent PR will add the analyzer logic to resolve SQL UDFs.

### Why are the changes needed?

To support SQL UDFs in Spark.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Existing tests. End to end tests will be added in the next PR once we support SQL UDF resolution.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #49389 from allisonwang-db/spark-50734-sql-udf-catalog-api.

Authored-by: Allison Wang <allison.wang@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>