[SPARK-50705][SQL] Make `QueryPlan` lock-free
### What changes were proposed in this pull request?
Replace a group of `lazy val` in `QueryPlan` with new lock-free helper classes.
Not all `lazy val`s are replaced in this PR, we will need to handle remaining `lazy val`s together with its subclasses to make it take effect.

### Why are the changes needed?
for the deadlock issues on query plan nodes:

- sometimes we want the plan node methods to use a coarse lock (just lock the plan node itself), as these methods (expressions , output , references , deterministic , schema , canonicalized , etc.) may call each other, so using a coarse lock can prevent deadlocks.
- sometimes we want these methods to use fine-grained locks, because these methods may call each other of a parent/child plan node. If you traverse the tree with different directions at the same time, it's likely to hit deadlock using coarse lock.

the only solution is to not use locks if possible

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Added unit tests
Manually test against a deadlock case

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #49212 from zhengruifeng/query_plan_atom_refs.

Lead-authored-by: Ruifeng Zheng <ruifengz@apache.org>
Co-authored-by: Josh Rosen <joshrosen@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>