[SPARK-50600][CONNECT][SQL] Set analyzed on analysis failure
### What changes were proposed in this pull request?
As part of [SPARK-44145](https://issues.apache.org/jira/browse/SPARK-44145), a callback was added to track completion of analysis and optimization phase of a query. While the analyzed plan is sent when analysis completes successfully it does not when it fail. In that case, we should fallback to the ParsedPlan.

### Why are the changes needed?
The purpose of the analyze event is to track when analysis completes, as such it should also be sent on both success & failure.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Unit

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #49383 from jdesjean/jdesjean/SPARK-50600.

Authored-by: jdesjean <jf.desjeans.gauthier@gmail.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>