[SPARK-50779][SQL] Adding feature flag for object level collations
### What changes were proposed in this pull request?

As a follow up from https://github.com/apache/spark/pull/49084 and associated JIRA issue https://issues.apache.org/jira/browse/SPARK-50675, adding an internal feature SQL conf flag around object level collations, to be able to enable/disable the feature.

### Why are the changes needed?

The object level collations feature is still in development phase, and so it should be disabled by default until it is completed, other than in unit tests that validate the functionality.

### Does this PR introduce _any_ user-facing change?

Just explicitly throwing a query compilation error that object level collations are unsupported feature, in case the newly added feature flag is disabled.

### How was this patch tested?

Dedicated unit tests are added to verify that when the feature flag is disabled, attempting to set object level collation (on table or view object) results in query compilation error that the feature is not supported. The existing unit tests already added as part of previous PR linked above validate that when the feature flag is enabled, the object level collation could be properly set (flag is enabled by default in testing environment)

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #49431 from dejankrak-db/conf-flag-object-level-collations.

Authored-by: Dejan Krakovic <dejan.krakovic@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>