[SPARK-50124][SQL][FOLLOWUP] InsertSortForLimitAndOffset should propagate missing ordering columns
### What changes were proposed in this pull request?

This is a followup of https://github.com/apache/spark/pull/48661 to extend the fix to handle a common query pattern: doing project/filter after the sort.

The implementation is a bit complicated as now we need to propagate the ordering columns pre-shuffle, so that we can perform the local sort post-shuffle.

### Why are the changes needed?

extend the bug fix to cover more query patterns.

### Does this PR introduce _any_ user-facing change?

No, the fix is not released

### How was this patch tested?

new tests

### Was this patch authored or co-authored using generative AI tooling?

no

Closes #49416 from cloud-fan/fix.

Authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>