[SPARK-50814][DOCS] Remove unused SQL error pages
### What changes were proposed in this pull request?

Remove standalone SQL error pages that were made obsolete by the work completed in #44971.

Also fix the formatting of the error message for `QUERY_ONLY_CORRUPT_RECORD_COLUMN`, since it was incorrect and overflowing the table cell it belongs to.

### Why are the changes needed?

These error pages are either already captured completely in `common/utils/src/main/resources/error/error-conditions.json`, or are obsolete and not needed (and are not being rendered in the documentation output anyway).

The formatting of `QUERY_ONLY_CORRUPT_RECORD_COLUMN` before and after:

<img src="https://github.com/user-attachments/assets/476c57e0-dfa5-403e-8a7d-2d05301eb7a3" width=650 />
<img src="https://github.com/user-attachments/assets/106d5bca-6569-488c-9b9c-1a27345fc7a8" width=450 />

### Does this PR introduce _any_ user-facing change?

Yes, documentation formatting.

### How was this patch tested?

Built the docs locally and reviewed them in my browser.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #49486 from nchammas/SPARK-50814-unused-error-docs.

Authored-by: Nicholas Chammas <nicholas.chammas@gmail.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>