[SPARK-50830][SQL] Return single-pass result as the dual run analysis result
### What changes were proposed in this pull request?
Right now fixed-point resolved plan is return as a result of analysis of the unresolved plan. In this task we add a flag in order to guard returning of the single-pass result in specific cases so it can be reverted more easily + return single-pass resolver result as the result of the dual-run analysis.

### Why are the changes needed?
Goal is to have single-pass resolver result as the global one (analysis result of an unresolved plan). We are introducing it now but also adding a flag so it can be reverted easily.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Existing tests with the `ANALYZER_DUAL_RUN_LEGACY_AND_SINGLE_PASS_RESOLVER` value set to true.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #49509 from mihailoale-db/mihailo-aleksic_data/mihailoale-db/singlepassresultflag.

Authored-by: mihailoale-db <mihailo.aleksic@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>