[SPARK-50885][BUILD] Change to use `SbtPomKeys` to obtain `io.grpc.version` in `SparkBuild.scala`
### What changes were proposed in this pull request?
This pr chanage to use

```
SbtPomKeys.effectivePom.value.getProperties.get("io.grpc.version").asInstanceOf[String]
```

to get `io.grpc.version`, so that when upgrading gRPC in the future, there will be no need to modify both `pom.xml` and `SparkBuild.scala` simultaneously.

On the other hand, this PR moves the version retrieval from `BuildCommons` to `SparkConnectCommon`, because currently only `SparkConnectCommon` requires the definition of this version.

### Why are the changes needed?
Simplify the management of the `grpcVersion`.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
- Pass GitHub Actions

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #49562 from LuciferYang/SPARK-50885.

Authored-by: yangjie01 <yangjie01@baidu.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>