[SPARK-50872][SQL][UI] Makes ToPrettyString expression not affect UI presentation
### What changes were proposed in this pull request?

The Plan graph and plan detail are affected by the injected ToPrettyString expression. They are not consistent with the output of our `explain` API. If there are a lot of output columns, the UI looks malformed.
![image](https://github.com/user-attachments/assets/f7f32369-b5c5-47d3-b320-ffaff55867a6)

### Why are the changes needed?

It's not good to keep noise on the user's daily debug tool

### Does this PR introduce _any_ user-facing change?

No, this only changes the ongoing 4.0

### How was this patch tested?

Added UT and UI manual test

![image](https://github.com/user-attachments/assets/f851096a-b952-419e-9062-3995e6b1697c)

### Was this patch authored or co-authored using generative AI tooling?

no

Closes #49551 from yaooqinn/SPARK-50872.

Authored-by: Kent Yao <yao@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>