[SPARK-49273][CONNECT][SQL] Origin support for Spark Connect Scala client
This PR adds support for `Origin` to the Scala client of Spark Connect.

The approach is to send the client `Origin` instance over the wire to the Server, which will reconstruct the instance and attach it to the expression during transformation. By default, only the last element in the stack trace is sent to the server. This is in line with the Python implementation and can be controlled via a Spark config `spark.sql.stackTracesInDataFrameContext`.

One Origin message can contain only one of `PythonOrigin` and `JvmOrigin`. The former is filled by the Python client (https://github.com/apache/spark/pull/46789) and the latter by the Scala client (this PR).

Close the gap between Classic & Connect.

No.

New tests.

Yes, to generate protobuf messages.

Closes #49373 from xupefei/connect-origin.

Authored-by: Paddy Xu <xupaddy@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>
(cherry picked from commit 0104f3843794575fa1a143061eaac1a4c75952a0)
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>