[SPARK-50847][SQL] Deny ApplyCharTypePadding from applying on specific In expressions
### What changes were proposed in this pull request?
`ApplyCharTypePadding` rule shouldn't be applied to `In` expressions where elements of the `In.list` are not `StringType`s and in this PR we fix it.

### Why are the changes needed?
Users should get a proper `DATATYPE_MISMATCH.DATA_DIFF_TYPES` error message instead of an `java.lang.ClassCastException` internal one.

### Does this PR introduce _any_ user-facing change?
Users will now get a proper message instead of an internal one.

### How was this patch tested?
Added test.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #49526 from mihailoale-db/mihailo-aleksic_data/fixinpadding.

Authored-by: mihailoale-db <mihailo.aleksic@databricks.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>