[SPARK-50735][CONNECT] Failure in ExecuteResponseObserver results in infinite reattaching requests
### What changes were proposed in this pull request?

The Spark Connect reattach request handler checks whether the associated ExecuteThreadRunner is completed, and returns an error if it has failed to record the outcome.

### Why are the changes needed?

ExecuteResponseObserver.{onError, onComplete} are fallible while they are not retried; this leads to a situation where the ExecuteThreadRunner is completed without succeeding in responding to the client, and thus the client keeps retrying by reattaching the execution.

To be specific, if an ExecuteThreadRunner fails to record the completion of execution or an error on the observer and then just disappears, the client will endlessly reattach, hoping that "someone" will eventually record "some data," but since the ExecuteThreadRunner is gone, this becomes a deadlock situation.

The fix is that when the client attaches, the handler checks the status of the ExecuteThreadRunner, and if it finds that the execution cannot make any progress, an error is returned.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

testOnly org.apache.spark.sql.connect.service.SparkConnectServiceE2ESuite

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #49370 from changgyoopark-db/SPARK-50735.

Lead-authored-by: changgyoopark-db <changgyoo.park@databricks.com>
Co-authored-by: Changgyoo Park <164183463+changgyoopark-db@users.noreply.github.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>
(cherry picked from commit 431aef8b35050510d3f90b99446568142c56a2fc)
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>