[SPARK-50897][ML][CONNECT] Avoiding instance creation in ServiceLoader
### What changes were proposed in this pull request?
When converting the iterator of the ServiceLoader call to Scala, we explicitly instantiate all classes that the service loader provides. Since we do not need the instances, this PR uses the `steam()` of the ServiceLoader to iterate over the list of providers and just extracts the clasess.

### Why are the changes needed?
Performance / Stability

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Existing tests

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #49577 from grundprinzip/spark_ml_service_loader.

Authored-by: Martin Grund <martin.grund@databricks.com>
Signed-off-by: Martin Grund <martin.grund@databricks.com>